modularization: extract utils
This required the options.show_branch flag to be extracted. It was
previously used as a global variable to control the mode of
`CoverageData.coverage()`.
